Split up Home Content component
